+++ ChangeLog.md
@@ -3,7 +3,7 @@
 ## v0.1.6
 
 1. 修复广发证券现金current_amount 取值错误
-2. 增加session 的reset方法，用于重置一下session
+2. 增加loginsession 的reset方法，用于重置一下session
 
 ## v0.1.5